Add missing TypeScript package and update readme
Unfortunately we have to keep a second copy for now. The lockfile
specifies the same version as used in the root package to avoid issues.
This may be replaced with Node.js type stripping in the future.

README.md is updated to include the correct command for testing the
Electron wrapper.