Handle directories in mod protocol handler (#87)
* Handle directories in mod protocol handler
Previously we would return an error about a file not existing, which kinda makes sense, considering a directory isn't a file.
Now with this commit, if the url resolve to a directory, it will return a json array with the relative paths of the file in that directory.

* Make directory index work with ASAR archives

Also fix directory index having higher priority than normal requests and
simplify URL handling.

---------

Co-authored-by: Даниїл Григор'єв <dengr1065@gmail.com>