Add draw hook for wires foreground layer
Without this change a runBeforeMethod/runAfterMethod is required to draw
anything on the wires layer, leading to some confusion and asymmetry
between the regular and wires layer.

Therefore, introduce a set of before/after hooks for the wires
foreground layer, allowing mods to easily draw stuff on top of or behind
wires layer components.